Update Start Fragment 01

+++ AdventureMaps/app/src/main/java/com/example/adventuremaps/Fragments/FragmentStartLocalizationPointClick.java
@@ -6,6 +6,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.LinearLayout;
+import android.widget.Toast;
 
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProviders;
@@ -26,14 +27,7 @@ public class FragmentStartLocalizationPointClick extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.fragment_click_localization_point, container, false);
 
-        /*LinearLayout.LayoutParams param = new LinearLayout.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                0,
-                (float) 1.0
-        );
-        param.weight = 30;
-        view.setLayoutParams(param);*/
-
+        //Instanciamos el VM
         viewModel = ViewModelProviders.of(getActivity()).get(MainTabbetActivityVM.class);
 
         btnDelete = view.findViewById(R.id.btnDeleteLocalizationPointFrgamentStart);
@@ -42,7 +36,11 @@ public class FragmentStartLocalizationPointClick extends Fragment {
         btnDelete.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                viewModel.deleteLocalizationDialog(getActivity());
+                if(viewModel.getLocalizationPoint().getEmailCreator().equals(viewModel.get_actualEmailUser())){
+                    viewModel.deleteLocalizationDialog(getActivity());
+                }else{
+                    Toast.makeText(getContext(), R.string.cant_delete_localization_owner, Toast.LENGTH_LONG).show();
+                }
             }
         });
